+++ lib/livekit_client.dart
@@ -45,7 +45,7 @@ export 'src/track/remote/remote.dart';
 export 'src/track/remote/video.dart';
 export 'src/track/track.dart';
 export 'src/track/processor.dart';
-export 'src/track/visualizer.dart';
+export 'src/track/audio_visualizer.dart';
 export 'src/types/other.dart';
 export 'src/types/participant_permissions.dart';
 export 'src/types/video_dimensions.dart';
