+++ lib/src/track/audio_visualizer.dart
